Change log index type from int to long

+++ src/main/java/alien4cloud/paas/yorc/context/rest/response/LogEventDTO.java
@@ -11,5 +11,5 @@ import java.util.List;
 @Setter
 public class LogEventDTO {
     private List<LogEvent> logs;
-    private int last_index;
+    private long last_index;
 }
